UserVerification/requirements.txt
py-cord==2.3.2
python-dotenv
flask